Make it possible to use app creds in accounting as well
