+++ src/fedcloud_catchall/test_discovery.py
@@ -51,6 +51,60 @@ SITES_INFO = [
     },
 ]
 
+LOADED_SITE = {
+    "hostname": "example.com",
+    "id": "1",
+    "name": "TEST",
+    "projects": [
+        {
+            "id": "abc",
+            "name": "ops",
+        },
+        {
+            "id": "def",
+            "name": "fake-vo.example.com",
+        },
+    ],
+    "shares": {
+        "fake-vo.example.com": {
+            "auth": {},
+            "id": "def",
+            "name": "fake-vo.example.com",
+        },
+        "ops": {
+            "auth": {
+                "somekey": "value",
+            },
+            "id": "abc",
+            "name": "ops",
+        },
+    },
+    "state": "",
+    "static": {
+        "endpoint": "https://example.com:5000/v3",
+        "gocdb": "TEST",
+        "images": {
+            "formats": [
+                "qcow2",
+                "raw",
+            ],
+            "sync": True,
+        },
+        "vos": [
+            {
+                "auth": {
+                    "somekey": "value",
+                },
+                "name": "ops",
+            },
+            {
+                "name": "fake-vo.example.com",
+            },
+        ],
+    },
+    "url": "https://example.com:5000/v3",
+}
+
 
 class TestDiscovery(testtools.TestCase):
     def setUp(self):
@@ -77,61 +131,7 @@ class TestDiscovery(testtools.TestCase):
         m_glob.return_value = ["file1.yaml"]
         with patch("builtins.open", mock_open(read_data=SAMPLE_SITE)):
             sites = disco.load_sites()
-            assert sites == {
-                "1": {
-                    "hostname": "example.com",
-                    "id": "1",
-                    "name": "TEST",
-                    "projects": [
-                        {
-                            "id": "abc",
-                            "name": "ops",
-                        },
-                        {
-                            "id": "def",
-                            "name": "fake-vo.example.com",
-                        },
-                    ],
-                    "shares": {
-                        "fake-vo.example.com": {
-                            "auth": {},
-                            "id": "def",
-                            "name": "fake-vo.example.com",
-                        },
-                        "ops": {
-                            "auth": {
-                                "somekey": "value",
-                            },
-                            "id": "abc",
-                            "name": "ops",
-                        },
-                    },
-                    "state": "",
-                    "static": {
-                        "endpoint": "https://example.com:5000/v3",
-                        "gocdb": "TEST",
-                        "images": {
-                            "formats": [
-                                "qcow2",
-                                "raw",
-                            ],
-                            "sync": True,
-                        },
-                        "vos": [
-                            {
-                                "auth": {
-                                    "somekey": "value",
-                                },
-                                "name": "ops",
-                            },
-                            {
-                                "name": "fake-vo.example.com",
-                            },
-                        ],
-                    },
-                    "url": "https://example.com:5000/v3",
-                },
-            }
+            assert sites == {"1": LOADED_SITE}
 
     @patch("fedcloud_catchall.discovery.fetch_site_info")
     @patch("glob.iglob")
@@ -167,6 +167,42 @@ class TestDiscovery(testtools.TestCase):
             path="users/user@egi.eu/cloudmon/example.com/ops",
             mount_point="/secrets/",
         )
+
+    def test_auth_config_oidc(self):
+        site = LOADED_SITE
+        config = [
+            s.strip()
+            for s in disco.auth_config(site, site["shares"]["ops"], "section").split(
+                "\n"
+            )
+        ]
+        assert "[section]" in config
+        assert "project_id = abc" in config
+        assert "access_token_type = access_token" in config
+        assert "auth_url = https://example.com:5000/v3" in config
+
+    @patch("fedcloud_catchall.discovery.generate_token")
+    @patch("fedcloud_catchall.discovery.get_oidc_config")
+    @patch("fedcloud_catchall.discovery.get_vo_secrets")
+    def test_auth_config_secret(self, m_secrets, m_oidc, m_token):
+        site = copy.deepcopy(LOADED_SITE)
+        site["static"]["auth"] = "v3applicationcredential"
+        m_secrets.return_value = {"abc": "foo"}
+        config = [
+            s.strip()
+            for s in disco.auth_config(site, site["shares"]["ops"], "section").split(
+                "\n"
+            )
+        ]
+        assert "[section]" in config
+        assert "auth_type = v3applicationcredential" in config
+        assert "auth_url = https://example.com:5000/v3" in config
+        assert "abc = foo" in config
+        assert "somekey = value" in config
+        # test that second call does not recall the token stuff
+        disco.auth_config(site, site["shares"]["ops"], "section2")
+        m_oidc.assert_called_once()
+        m_token.assert_called_once()
 
 
 if __name__ == "__main__":
